Resolved conflicts after changes
